api/requirements.txt
fastapi>=0.128.0,<1.0.0
uvicorn[standard]>=0.40.0,<1.0.0
sqlalchemy>=2.0.46,<3.0.0
alembic>=1.18.3,<2.0.0
psycopg2-binary>=2.9.11,<3.0.0
redis>=7.1.0,<8.0.0
pyjwt>=2.11.0,<3.0.0
opentelemetry-api>=1.39.1,<2.0.0
opentelemetry-sdk>=1.39.1,<2.0.0
opentelemetry-exporter-otlp>=1.39.1,<2.0.0
opentelemetry-instrumentation-fastapi>=0.60b1,<1.0.0
opentelemetry-instrumentation-redis>=0.60b1,<1.0.0
opentelemetry-instrumentation-sqlalchemy>=0.60b1,<1.0.0
pytest>=9.0.2,<10.0.0
pytest-cov>=5.0.0,<6.0.0
httpx>=0.28.1,<1.0.0
boto3>=1.35.0,<2.0.0
